backend/pyproject.toml
[project]
name = "applied-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "depthai>=3.2.1",
    "fastapi>=0.123.5",
    "google-genai>=1.52.0",
    "opencv-python>=4.12.0.88",
    "python-dotenv>=1.2.1",
    "ultralytics>=8.3.234",
    "uvicorn>=0.38.0",
]
